Fix/support callback for define()
The define() function returns a function which has supported
an input parameter and was supposed to support a callback. The
callback mechanism was not working correctly and was only
running define's first function, so now it's fixed.

The test has been adapted to ensure that multiple functions
are chained before verifying that the callback succeeds.

Signed-off-by: Eric Windisch <eric@windisch.us>